Move test classes to the correct package.
